+++ R/prepare_annotations_gnps.R
@@ -159,7 +159,7 @@ prepare_annotations_gnps <- function(
     parameters = get_params(step = "prepare_annotations_gnps"),
     step = "prepare_annotations_gnps"
   )
-  export_output(x = table, file = output[[1]])
+  export_output(x = table, file = output[[1L]])
   rm(table)
-  return(output[[1]])
+  return(output[[1L]])
 }